Replace ProcessDrawer with ProcessDropdown
