+++ src/lib/components/processing/stories/ProcessDropdown.stories.svelte
@@ -1,14 +1,14 @@
 <script context="module" lang="ts">
   import { Story } from "@storybook/addon-svelte-csf";
   import ProcessContext from "../ProcessContext.svelte";
-  import ProcessDrawer from "../ProcessDrawer.svelte";
+  import ProcessDropdown from "../ProcessDropdown.svelte";
 
   import { runs } from "@/test/handlers/addons";
   import * as mock from "@/test/handlers/documents";
 
   export const meta = {
-    title: "Components / Processing / Drawer",
-    component: ProcessDrawer,
+    title: "Components / Processing / Dropdown",
+    component: ProcessDropdown,
     parameters: { layout: "fullscreen" },
   };
 </script>
@@ -22,6 +22,6 @@
   }}
 >
   <ProcessContext>
-    <ProcessDrawer />
+    <ProcessDropdown />
   </ProcessContext>
 </Story>
